Dropping old interpreter
