+++ .eslintrc
@@ -103,7 +103,18 @@
     "require-yield": "off",
     "no-case-declarations": "off",
     "no-console": [
-      "warn"
+      "warn",
+      {
+        "allow": [
+          "info",
+          "warn",
+          "error",
+          "time",
+          "timeEnd",
+          "group",
+          "groupEnd"
+        ]
+      }
     ],
     "no-constant-condition": [
       "warn"
