+++ package.json
@@ -19,10 +19,8 @@
     "test:unit": "mocha --parallel -r ts-node/register test/**/*.test.ts",
     "test:examples": "npm run test:wtest -- --root language/test/examples",
     "test:sanity": "npm run test:wtest -- --root language/test/sanity",
-    "test:sanity2": "npm run test:wtest2 -- --root language/test/sanity",
     "test:game": "npm run test:wtest -- --root test/game",
     "test:wtest": "mocha --delay -t 10000 -r ts-node/register test/wtest.ts",
-    "test:wtest2": "mocha --delay -t 10000 -r ts-node/register test/wtest2.ts",
     "prepublishOnly": "npm run build && npm test",
     "postpublish": "git tag v$npm_package_version && git push --tags",
     "prepack": "npm run build",
