+++ src/interpreter/runtimeModel.ts
@@ -1,539 +1,101 @@
-import { last, get } from '../extensions'
-import { Node, Environment, Id, List, Module, Name, Variable, Singleton, isNode, Method } from '../model'
+import { Environment, is, isNode, Method, Module, Name, Node, Variable, Singleton, Expression, List, Class, Id } from '../model'
+import { get, last } from '../extensions'
 import { v4 as uuid } from 'uuid'
-import { Logger, nullLogger } from './log'
-import compile, { PUSH, INIT, Instruction, NULL_ID, TRUE_ID, FALSE_ID } from './compiler'
-import takeStep from './interpreter'
 
 const { isArray } = Array
-const { assign, keys } = Object
-
-export type Locals = Map<Name, RuntimeObject | LazyInitializer | undefined>
-export type NativeFunction = (self: RuntimeObject, ...args: RuntimeObject[]) => (evaluation: Evaluation) => void
-export interface Natives { [name: string]: NativeFunction | Natives }
-
+const { keys, entries } = Object
 
 const DECIMAL_PRECISION = 5
-const MAX_FRAME_STACK_SIZE = 1000
-const MAX_OPERAND_STACK_SIZE = 10000
 
-
-const COPY = Symbol('copy')
-const CREATE_OR_RETRIEVE = Symbol('createOrRetrieve')
-
-// ══════════════════════════════════════════════════════════════════════════════════════════════════════════════════
-// ERRORS
-// ══════════════════════════════════════════════════════════════════════════════════════════════════════════════════
-
-export class WollokError extends Error { constructor(readonly moduleFQN: Name) { super(moduleFQN) } }
-
-export class WollokUnrecoverableError extends Error { }
-
-// ══════════════════════════════════════════════════════════════════════════════════════════════════════════════════
-// EVALUATION
-// ══════════════════════════════════════════════════════════════════════════════════════════════════════════════════
-
-export class Evaluation {
-  log: Logger
-  readonly environment: Environment
-  readonly rootContext: Context
-  readonly frameStack: Stack<Frame>
-  readonly natives: Natives
-  protected readonly instanceCache: Map<Id, RuntimeObject>
-
-  get currentFrame(): Frame | undefined { return this.frameStack.top }
-  get currentContext(): Context { return this.currentFrame?.context ?? this.rootContext }
-  get instances(): List<RuntimeObject> { return [...this.instanceCache.values()] }
-
-
-  static create(environment: Environment, natives: Natives, stepAllInitialization = true): Evaluation {
-    const rootContext = new Context()
-    const evaluation = new Evaluation(environment, natives, () => rootContext)
-
-    const globalConstants = environment.descendants().filter((node: Node): node is Variable => node.is('Variable') && node.parent().is('Package'))
-    const globalSingletons = environment.descendants().filter((node: Node): node is Singleton => node.is('Singleton') && !!node.name)
-
-    rootContext.set('null', RuntimeObject.null(evaluation))
-    rootContext.set('true', RuntimeObject.boolean(evaluation, true))
-    rootContext.set('false', RuntimeObject.boolean(evaluation, false))
-
-    for (const module of globalSingletons)
-      rootContext.set(module.fullyQualifiedName(), RuntimeObject.object(evaluation, module))
-
-    for (const constant of globalConstants)
-      rootContext.set(constant.fullyQualifiedName(), new LazyInitializer(evaluation, rootContext, constant.fullyQualifiedName(), compile(constant.value)))
-
-    evaluation.pushFrame(new Frame(rootContext, [
-      ...globalSingletons.flatMap(singleton => {
-        const supercallArgs = singleton.supertypes.flatMap(supertype => supertype.args)
-        return [
-          ...supercallArgs.flatMap(({ value }) => compile(value)),
-          PUSH(singleton.id),
-          INIT(supercallArgs.map(({ name }) => name)),
-        ]
-      }),
-    ]))
-
-    if (stepAllInitialization) {
-      evaluation.log.start('Initializing Evaluation')
-      evaluation.stepAll()
-      evaluation.log.done('Initializing Evaluation')
-    }
-
-    return evaluation
-  }
-
-
-  protected constructor(
-    environment: Environment,
-    natives: Natives,
-    rootContext: (evaluation: Evaluation) => Context,
-    frameStack = (_evaluation: Evaluation) => new Stack<Frame>(MAX_FRAME_STACK_SIZE),
-    instanceCache = (_evaluation: Evaluation) => new Map<Id, RuntimeObject>(),
-    logger: Logger = nullLogger,
-  ) {
-    this.environment = environment
-    this.natives = natives
-    this.rootContext = rootContext(this)
-    this.frameStack = frameStack(this)
-    this.instanceCache = instanceCache(this)
-    this.log = logger
-  }
-
-
-  [CREATE_OR_RETRIEVE](instance: RuntimeObject): RuntimeObject {
-    const existing = this.instanceCache.get(instance.id)
-    if (existing) return existing
-
-    this.instanceCache.set(instance.id, instance)
-
-    return instance
-  }
-
-  copy(): Evaluation {
-    const cache = new Map<Id, any>()
-
-    return new Evaluation(
-      this.environment,
-      this.natives,
-      evaluation => this.rootContext[COPY](evaluation, cache),
-      evaluation => this.frameStack.map(frame => frame[COPY](evaluation, cache)),
-      evaluation => new Map([...this.instanceCache.entries()].map(([name, instance]) => [name, instance[COPY](evaluation, cache)])),
-      this.log,
-    )
-  }
-
-  pushFrame(frame: Frame): void { this.frameStack.push(frame) }
-
-  popFrame(): void { this.frameStack.pop() }
-
-  instance(id: Id): RuntimeObject {
-    const response = this.instanceCache.get(id)
-    if (!response) throw new RangeError(`Access to undefined instance "${id}"`)
-    return response
-  }
-
-  freeInstance(id: Id): void { this.instanceCache.delete(id) }
-
-  invoke(methodOrMessage: Method | string, receiver: RuntimeObject, ...args: RuntimeObject[]): void {
-    const method = methodOrMessage instanceof Method ? methodOrMessage : receiver.module.lookupMethod(methodOrMessage, args.length)
-    if (!method) throw new Error(`Can't invoke unexistent method ${receiver.module.fullyQualifiedName()}.${methodOrMessage}/${args.length}`)
-    if (method.isAbstract()) throw new Error(`Can't invoke abstract method ${method.parent().fullyQualifiedName()}.${method.name}/${method.parameters.length}`)
-    if (methodOrMessage instanceof Method && !method.matchesSignature(method.name, args.length)) throw new Error(`Wrong number of arguments (${args.length}) for method ${method.parent().fullyQualifiedName()}.${method.name}/${method.parameters.length}`)
-
-    if (method.body === 'native') {
-      const nativeFQN = `${method.parent().fullyQualifiedName()}.${method.name}`
-      const native = get<NativeFunction>(this.natives, nativeFQN)
-      if (native) this.log.debug('Invoking Native:', nativeFQN, '/', args.length)
-      else throw new Error(`Native not found: ${nativeFQN}`)
-      native(receiver, ...args)(this)
-    } else {
-      this.pushFrame(new Frame(
-        receiver,
-        compile(method),
-        new Map(method.parameters.map(({ name, isVarArg }, index) => [name,
-          isVarArg
-            ? RuntimeObject.list(this, args.slice(index).map(({ id }) => id))
-            : args[index],
-        ]))
-      ))
-    }
-  }
-
-  raise(exception: RuntimeObject): void {
-    while (!this.frameStack.isEmpty()) {
-      const currentFrame = this.currentFrame!
-
-      while (currentFrame.hasNestedContext()) {
-        if (currentFrame.context.exceptionHandlerIndex !== undefined) {
-          currentFrame.jumpTo(currentFrame.context.exceptionHandlerIndex)
-          currentFrame.popContext()
-          currentFrame.pushOperand(exception)
-          return
-        }
-
-        currentFrame.popContext()
-      }
-
-      this.popFrame()
-    }
-
-    throw new WollokUnrecoverableError(`Reached end of stack with unhandled exception ${exception.module.fullyQualifiedName()}: ${exception.get('message')?.innerValue}`)
-  }
-
-  stepIn(): void {
-    this.log.step(this)
-    takeStep(this)
-  }
-
-  stepOut(): void {
-    const initialFrameDepth = this.frameStack.depth
-    do {
-      this.stepIn()
-    } while (this.frameStack.depth >= initialFrameDepth)
-  }
-
-  stepOver(): void {
-    const initialFrameDepth = this.frameStack.depth
-    do {
-      this.stepIn()
-    } while (this.frameStack.depth > initialFrameDepth)
-  }
-
-  /** Takes all possible steps, until the last frame has no pending instructions and then drops that frame */
-  stepAll(): void {
-    while (!this.currentFrame!.isFinished()) this.stepIn()
-    this.popFrame()
-  }
-
-  // TODO: stepThrough
-}
-
-// ══════════════════════════════════════════════════════════════════════════════════════════════════════════════════
-// STACKS
-// ══════════════════════════════════════════════════════════════════════════════════════════════════════════════════
-
-export class Stack<T> {
-  protected readonly maxSize: number
-  protected readonly elements: T[] = []
-
-  constructor(maxSize: number) {
-    this.maxSize = maxSize
-  }
-
-  get depth(): number { return this.elements.length }
-
-  get top(): T | undefined { return last(this.elements) }
-
-  *[Symbol.iterator](): Iterator<T> { yield* this.elements }
-
-  isEmpty(): boolean { return !this.elements.length }
-
-  forEach(f: (element: T) => void): void { this.elements.forEach(f) }
-
-  map<U>(tx: (element: T) => U): Stack<U> {
-    const response = new Stack<U>(this.maxSize)
-    response.unshift(...this.elements.map(tx))
-    return response
-  }
-
-  unshift(...elements: T[]): void {
-    if (this.maxSize! < this.elements.length + elements.length)
-      throw new WollokError('wollok.lang.StackOverflowException')
-
-    this.elements.unshift(...elements)
-  }
-
-  push(...elements: T[]): void {
-    if (this.maxSize! < this.elements.length + elements.length)
-      throw new WollokError('wollok.lang.StackOverflowException')
-
-    this.elements.push(...elements)
-  }
-
-  pop(): T {
-    if (!this.elements.length) throw new Error('Stack underflow')
-    return this.elements.pop()!
-  }
-}
-
-// ══════════════════════════════════════════════════════════════════════════════════════════════════════════════════
-// FRAMES
-// ══════════════════════════════════════════════════════════════════════════════════════════════════════════════════
-
-export class Frame {
-  readonly operandStack = new Stack<RuntimeObject | undefined>(MAX_OPERAND_STACK_SIZE)
-  readonly instructions: List<Instruction>
-  readonly baseContext: Context
-  protected pc = 0
-  protected currentContext: Context
-
-  get nextInstructionIndex(): number { return this.pc }
-  get context(): Context { return this.currentContext }
-
-
-  [COPY](evaluation: Evaluation, cache: Map<Id, any>): Frame {
-    const copy = new Frame(this.context[COPY](evaluation, cache), this.instructions)
-
-    assign(copy, { currentContext: copy.context.parentContext, initialContext: copy.context.parentContext })
-    copy.operandStack.unshift(...this.operandStack.map(operand => operand?.[COPY](evaluation, cache)))
-    copy.pc = this.pc
-
-    return copy
-  }
-
-
-  constructor(parentContext: Context, instructions: List<Instruction>, locals: Locals = new Map()) {
-    this.baseContext = new Context(parentContext)
-    this.currentContext = this.baseContext
-    this.instructions = instructions
-
-    locals.forEach((instance, name) => this.baseContext.set(name, instance))
-  }
-
-
-  isFinished(): boolean { return this.pc >= this.instructions.length }
-
-  hasNestedContext(): boolean { return this.context !== this.baseContext }
-
-  pushOperand(operand: RuntimeObject | undefined): void {
-    this.operandStack.push(operand)
-  }
-
-  popOperand(): RuntimeObject | undefined {
-    return this.operandStack.pop()
-  }
-
-  pushContext(exceptionHandlerIndex?: number): void {
-    this.currentContext = new Context(this.currentContext, exceptionHandlerIndex)
-  }
-
-  popContext(): void {
-    if (!this.hasNestedContext()) throw new Error('Popped frame base context')
-    this.currentContext = this.currentContext.parentContext!
-  }
-
-  takeNextInstruction(): Instruction {
-    if (this.isFinished()) throw new Error('Reached end of instructions')
-    return this.instructions[this.pc++]
-  }
-
-  jump(instructionDelta: number): void {
-    this.jumpTo(this.pc + instructionDelta)
-  }
-
-  jumpTo(instructionIndex: number): void {
-    if (instructionIndex < 0 || instructionIndex >= this.instructions.length)
-      throw new Error(`Out of range jump to instruction index ${instructionIndex}`)
-
-    this.pc = instructionIndex
-  }
-
-}
+export type NativeFunction = (this: Runner, self: RuntimeObject, ...args: RuntimeObject[]) => Generator<Node, RuntimeObject | undefined>
+export interface Natives { [name: string]: NativeFunction | Natives }
 
 // ══════════════════════════════════════════════════════════════════════════════════════════════════════════════════
 // CONTEXTS
 // ══════════════════════════════════════════════════════════════════════════════════════════════════════════════════
 
 export class Context {
-  readonly id: Id
+  readonly id: Id = uuid()
   readonly parentContext?: Context
-  readonly locals: Locals = new Map()
-  readonly exceptionHandlerIndex?: number
+  protected readonly locals: Map<Name, RuntimeObject | Generator<Node, RuntimeObject> | undefined> = new Map()
 
-  constructor(parent?: Context, exceptionHandlerIndex?: number, id: Id = uuid()) {
-    this.parentContext = parent
-    this.exceptionHandlerIndex = exceptionHandlerIndex
-    this.id = id
-  }
-
-  [COPY](evaluation: Evaluation, cache: Map<Id, any>): Context {
-    const cached = cache.get(this.id)
-    if (cached) return cached
-
-    const copy = new Context(
-      this.parentContext?.[COPY](evaluation, cache),
-      this.exceptionHandlerIndex,
-      this.id,
-    )
-
-    cache.set(this.id, copy)
-
-    this.locals.forEach((local, name) => copy.locals.set(name, local?.[COPY](evaluation, cache)))
-
-    return copy
+  constructor(parentContext?: Context, locals: Record<Name, RuntimeObject | Generator<Node, RuntimeObject>> = {}) {
+    this.parentContext = parentContext
+    for(const [name, value] of entries(locals)) this.locals.set(name, value)
   }
 
   get(local: Name): RuntimeObject | undefined {
-    return (this.locals.get(local) ?? this.parentContext?.get(local))?.materialize()
+    const found = this.locals.get(local) ?? this.parentContext?.get(local)
+    if (!found || found instanceof RuntimeObject) return found
+    let lazy = found.next()
+    while(!lazy.done) lazy = found.next()
+    this.set(local, lazy.value)
+    return lazy.value
   }
 
-  set(local: Name, value: RuntimeObject | LazyInitializer | undefined): void {
-    this.locals.set(local, value)
+  set(local: Name, value: RuntimeObject | Generator<Node, RuntimeObject>  | undefined, lookup = false): void {
+    if(!lookup || this.locals.has(local)) this.locals.set(local, value)
+    else this.parentContext?.set(local, value, lookup)
   }
 
-  contextHierarchy(): List<Context> {
-    return [this, ...this.parentContext?.contextHierarchy() ?? []]
-  }
-}
+  copy(contextCache: Map<Id, Context>): this {
+    if(contextCache.has(this.id)) return contextCache.get(this.id) as this
 
-// ══════════════════════════════════════════════════════════════════════════════════════════════════════════════════
-// RUNTIME OBJECTS
-// ══════════════════════════════════════════════════════════════════════════════════════════════════════════════════
+    const copy = this.baseCopy(contextCache) as this
+    contextCache.set(this.id, copy)
 
-export type InnerValue = string | number | Id[]
-
-
-export class LazyInitializer {
-  constructor(
-      protected readonly evaluation: Evaluation,
-      readonly context: Context,
-      readonly local: Name,
-      readonly instructions: List<Instruction>,
-  ) { }
-
-  [COPY](evaluation: Evaluation, cache: Map<Id, Context>): LazyInitializer {
-    return new LazyInitializer(evaluation, this.context[COPY](evaluation, cache), this.local, this.instructions)
-  }
-
-  materialize(): RuntimeObject | undefined {
-    this.evaluation.log.info('Materializing lazy initializer of', this.local)
-    const frame = new Frame(this.context, this.instructions)
-    this.evaluation.pushFrame(frame)
-    this.evaluation.stepAll()
-
-    const value = frame.popOperand()
-
-    this.evaluation.log.info('Materialized lazy initializer of', this.local, 'to', value?.id)
-
-    this.context.set(this.local, value)
-    return value
-  }
-}
-
-
-export class RuntimeObject extends Context {
-  readonly parentContext!: Context
-  readonly module: Module
-  readonly innerValue?: InnerValue
-
-  static null(evaluation: Evaluation): RuntimeObject {
-    return evaluation[CREATE_OR_RETRIEVE](
-      new RuntimeObject(
-        evaluation.currentContext,
-        evaluation.environment.getNodeByFQN('wollok.lang.Object'),
-        undefined,
-        NULL_ID,
+    for(const [name, value] of this.locals.entries())
+      copy.set(name,
+        value instanceof RuntimeObject ? value.copy(contextCache) :
+        value ? this.get(name) :
+        value
       )
-    )
-  }
-
-  static boolean(evaluation: Evaluation, value: boolean): RuntimeObject {
-    return evaluation[CREATE_OR_RETRIEVE](
-      new RuntimeObject(
-        evaluation.currentContext,
-        evaluation.environment.getNodeByFQN('wollok.lang.Boolean'),
-        undefined,
-        value ? TRUE_ID : FALSE_ID,
-      )
-    )
-  }
-
-  static number(evaluation: Evaluation, value: number): RuntimeObject {
-    const stringValue = value.toFixed(DECIMAL_PRECISION)
-    const id = `N!${stringValue}`
-
-    return evaluation[CREATE_OR_RETRIEVE](
-      new RuntimeObject(
-        evaluation.rootContext,
-        evaluation.environment.getNodeByFQN('wollok.lang.Number'),
-        Number(stringValue),
-        id,
-      )
-    )
-  }
-
-  static string(evaluation: Evaluation, value: string): RuntimeObject {
-    return evaluation[CREATE_OR_RETRIEVE](
-      new RuntimeObject(
-        evaluation.rootContext,
-        evaluation.environment.getNodeByFQN('wollok.lang.String'),
-        value,
-        `S!${value}`,
-      )
-    )
-  }
-
-  static list(evaluation: Evaluation, elements: List<Id>): RuntimeObject {
-    return evaluation[CREATE_OR_RETRIEVE](
-      new RuntimeObject(
-        evaluation.currentContext,
-        evaluation.environment.getNodeByFQN('wollok.lang.List'),
-        [...elements],
-      )
-    )
-  }
-
-  static set(evaluation: Evaluation, elements: List<Id>): RuntimeObject {
-    return evaluation[CREATE_OR_RETRIEVE](
-      new RuntimeObject(
-        evaluation.currentContext,
-        evaluation.environment.getNodeByFQN('wollok.lang.Set'),
-        [...elements],
-      )
-    )
-  }
-
-  static object(evaluation: Evaluation, moduleOrFQN: Module | Name, locals: Record<Name, RuntimeObject | undefined> = {}): RuntimeObject {
-    const module = isNode(moduleOrFQN) ? moduleOrFQN : evaluation.environment.getNodeByFQN<Module>(moduleOrFQN)
-    const instance = new RuntimeObject(
-      evaluation.currentContext,
-      module,
-      undefined,
-      module.is('Describe') || module.is('Singleton') && !!module.name ? module.id : undefined
-    )
-
-    for (const local of keys(locals)) instance.set(local, locals[local])
-
-    return evaluation[CREATE_OR_RETRIEVE](instance)
-  }
-
-  protected constructor(parent: Context, module: Module, innerValue?: InnerValue, id?: Id) {
-    super(parent, undefined, id)
-
-    this.module = module
-    this.innerValue = innerValue
-
-    this.locals.set('self', this)
-  }
-
-  [COPY](evaluation: Evaluation, cache: Map<Id, any>): RuntimeObject {
-    const cached = cache.get(this.id)
-    if (cached) return cached
-
-    const copy = new RuntimeObject(
-      this.parentContext[COPY](evaluation, cache),
-      this.module,
-      isArray(this.innerValue) ? [...this.innerValue] : this.innerValue,
-      this.id,
-    )
-
-    cache.set(this.id, copy)
-
-    this.locals.forEach((local, name) => copy.locals.set(name, local?.[COPY](evaluation, cache)))
 
     return copy
   }
 
-  materialize(): RuntimeObject | undefined { return this }
+  protected baseCopy(contextCache: Map<Id, Context>): Context {
+    return new Context(this.parentContext?.copy(contextCache))
+  }
+}
+
+
+export type InnerValue = boolean | string | number | RuntimeObject[] | Error
+
+export class RuntimeObject extends Context {
+  readonly module: Module
+  readonly innerValue?: InnerValue
+
+  constructor(module: Module, parentContext: Context, innerValue?: InnerValue) {
+    super(parentContext)
+    this.module = module
+    this.innerValue = innerValue
+    this.set('self', this)
+  }
+
+
+  protected baseCopy(contextCache: Map<Id, Context>): RuntimeObject {
+    return new RuntimeObject(
+      this.module,
+      this.parentContext!.copy(contextCache),
+      isArray(this.innerValue) ? this.innerValue.map(elem => elem.copy(contextCache)) : this.innerValue
+    )
+  }
+
+
+  assertIsException(): asserts this is RuntimeObject & { innerValue?: Error } {
+    if(this.innerValue && !(this.innerValue instanceof Error)) throw new TypeError('Malformed Runtime Object: Exception inner value, if defined, should be an Error')
+  }
+
+  assertIsBoolean(): asserts this is RuntimeObject & { innerValue: boolean } { this.assertIs('wollok.lang.Boolean', 'boolean') }
 
   assertIsNumber(): asserts this is RuntimeObject & { innerValue: number } { this.assertIs('wollok.lang.Number', 'number') }
 
   assertIsString(): asserts this is RuntimeObject & { innerValue: string } { this.assertIs('wollok.lang.String', 'string') }
 
-  assertIsCollection(): asserts this is RuntimeObject & { innerValue: Id[] } {
-    if (!isArray(this.innerValue) || this.innerValue.length && typeof this.innerValue[0] !== 'string')
-      throw new TypeError(`Malformed Runtime Object: Collection inner value should be a List<Id> but was ${this.innerValue}`)
+  assertIsCollection(): asserts this is RuntimeObject & { innerValue: RuntimeObject[] } {
+    if (!isArray(this.innerValue) || this.innerValue.length && !(this.innerValue[0] instanceof RuntimeObject))
+      throw new TypeError(`Malformed Runtime Object: Collection inner value should be a List<RuntimeObject> but was ${this.innerValue}`)
   }
 
   protected assertIs(moduleFQN: Name, innerValueType: string): void {
@@ -541,5 +103,390 @@ export class RuntimeObject extends Context {
       throw new TypeError(`Expected an instance of ${moduleFQN} but got a ${this.module.fullyQualifiedName()} instead`)
     if (typeof this.innerValue !== innerValueType)
       throw new TypeError(`Malformed Runtime Object: invalid inner value ${this.innerValue} for ${moduleFQN} instance`)
+  }
+}
+
+// ══════════════════════════════════════════════════════════════════════════════════════════════════════════════════
+// RUNNER CONTROLLER
+// ══════════════════════════════════════════════════════════════════════════════════════════════════════════════════
+
+class RunnerController {
+  generator: Generator<Node, RuntimeObject | undefined>
+  done = false
+
+  constructor(runner: Runner, node: Node) {
+    this.generator = runner.exec(node)
+  }
+
+  resume(breakpoints: Node[] = []): RuntimeObject | undefined {
+    if(this.done) throw new Error('Evaluation is already finished')
+
+    let next = this.generator.next()
+    while(!next.done) {
+      if(breakpoints.includes(next.value)) return
+      next = this.generator.next()
+    }
+
+    this.done = true
+    return next.value
+  }
+}
+
+// ══════════════════════════════════════════════════════════════════════════════════════════════════════════════════
+// RUNNER
+// ══════════════════════════════════════════════════════════════════════════════════════════════════════════════════
+
+export class WollokReturn extends Error { constructor(readonly instance?: RuntimeObject){ super() } }
+export class WollokException extends Error { constructor(readonly frameStack: List<Frame>, readonly instance: RuntimeObject){ super() } }
+
+
+export default class Interpreter {
+  readonly environment: Environment
+  readonly natives: Natives
+
+  constructor(environment: Environment, natives: Natives) {
+    this.environment = environment
+    this.natives = natives
+  }
+}
+
+
+export class Frame {
+  readonly node: Node
+  readonly context: Context
+
+  constructor(node: Node, context: Context) {
+    this.node = node
+    this.context = context
+  }
+
+  copy(contextCache: Map<Id, Context>): Frame {
+    return new Frame(this.node, this.context.copy(contextCache))
+  }
+}
+
+
+export class Runner {
+  readonly natives: Natives
+  readonly frameStack: Frame[]
+  console: Console = console
+
+  get currentContext(): Context { return last(this.frameStack)!.context }
+  get rootContext(): Context { return this.frameStack[0].context }
+  get environment(): Environment { return this.frameStack[0].node as Environment }
+
+  static build(environment: Environment, natives: Natives): Runner {
+    const evaluation = new Runner(natives, [new Frame(environment, new Context())])
+
+    evaluation.rootContext.set('null', evaluation.instantiate(environment.getNodeByFQN('wollok.lang.Object')))
+
+    const globalSingletons = environment.descendants().filter((node: Node): node is Singleton => node.is('Singleton') && !!node.name)
+    for (const module of globalSingletons)
+      evaluation.rootContext.set(module.fullyQualifiedName(), evaluation.instantiate(module))
+
+
+    const globalConstants = environment.descendants().filter((node: Node): node is Variable => node.is('Variable') && node.parent().is('Package'))
+    for (const constant of globalConstants)
+      evaluation.rootContext.set(constant.fullyQualifiedName(), evaluation.exec(constant.value, evaluation.rootContext))
+
+
+    for (const module of globalSingletons) {
+      const instance = evaluation.rootContext.get(module.fullyQualifiedName())
+      for (const field of module.defaultFieldValues().keys())
+        instance!.get(field.name)
+    }
+
+    for (const constant of globalConstants)
+      evaluation.rootContext.get(constant.fullyQualifiedName())
+
+
+    return evaluation
+  }
+
+  protected constructor(natives: Natives, frameStack: Frame[]) {
+    this.natives = natives
+    this.frameStack = frameStack
+  }
+
+  copy(contextCache: Map<Id, Context> = new Map()): Runner {
+    return new Runner(this.natives, this.frameStack.map(frame => frame.copy(contextCache)))
+  }
+
+  // ──────────────────────────────────────────────────────────────────────────────────────────────────────────────────
+  // EXECUTION
+  // ──────────────────────────────────────────────────────────────────────────────────────────────────────────────────
+
+  exec(node: Expression, context?: Context): Generator<Node, RuntimeObject>
+  exec(node: Node, context?: Context): Generator<Node, undefined>
+  *exec(node: Node, context: Context = this.currentContext): Generator<Node, RuntimeObject | undefined> {
+    this.frameStack.push(new Frame(node, context))
+
+    try {
+      if(node.is('Body')) {
+        yield node
+
+        let result: RuntimeObject | undefined
+        for(const sentence of node.sentences) result = yield* this.exec(sentence)
+        return result
+      }
+
+
+      if(node.is('Variable')) {
+        const value = yield* this.exec(node.value)
+        yield node
+        context.set(node.name, value)
+        return
+      }
+
+
+      if(node.is('Assignment')) {
+        const value = yield* this.exec(node.value)
+        yield node
+        if(node.variable.target()?.isReadOnly) throw new Error(`Can't assign the constant ${node.variable.target()?.name}`)
+        context.set(node.variable.name, value, true)
+        return
+      }
+
+
+      if(node.is('Return')) {
+        const value = node.value && (yield* this.exec(node.value))
+        yield node
+        throw new WollokReturn(value)
+      }
+
+
+      if(node.is('Reference')) {
+        yield node
+
+        const target = node.target()
+
+        return context.get(
+          target.is('Module') || target.is('Variable') && target.parent().is('Package')
+            ? target.fullyQualifiedName()
+            : node.name
+        )
+      }
+
+
+      if(node.is('Self')) {
+        yield node
+        return context.get('self')
+      }
+
+
+      if(node.is('Literal')) {
+        if(isArray(node.value)) {
+          const [reference, args] = node.value
+          const module = reference.target()!
+
+          const values: RuntimeObject[] = []
+          for(const arg of args) values.push(yield * this.exec(arg))
+
+          yield node
+
+          return yield* module.name === 'List' ? this.list(values) : this.set(values)
+        }
+
+        if (isNode(node.value)) {
+          yield node
+          return yield* this.instantiate(node.value)
+        }
+
+        yield node
+        return yield* this.reify(node.value as any)
+      }
+
+
+      if(node.is('New')) {
+        const args: Record<Name, RuntimeObject> = {}
+        for(const arg of node.args) args[arg.name] = yield* this.exec(arg.value)
+
+        yield node
+
+        if(!node.instantiated.target()) throw new Error(`Unexistent module ${node.instantiated.name}`)
+
+        return yield* this.instantiate(node.instantiated.target()!, args)
+      }
+
+
+      if(node.is('Send')) {
+        const receiver = yield* this.exec(node.receiver)
+        const values: RuntimeObject[] = []
+        for(const arg of node.args) values.push(yield * this.exec(arg))
+
+        yield node
+
+        return yield* this.invoke(node.message, receiver, ...values)
+      }
+
+
+      if(node.is('Super')) {
+        const values: RuntimeObject[] = []
+        for(const arg of node.args) values.push(yield * this.exec(arg))
+
+        yield node
+
+        const receiver = context.get('self')!
+        const currentMethod = node.ancestors().find(is('Method'))!
+        const method = receiver.module.lookupMethod(currentMethod.name, node.args.length, currentMethod.parent().fullyQualifiedName())
+        return yield* this.invoke(method, receiver, ...values)
+      }
+
+
+      if(node.is('If')) {
+        const condition: RuntimeObject = yield* this.exec(node.condition)
+        condition.assertIsBoolean()
+
+        yield node
+
+        return yield* this.exec(condition.innerValue ? node.thenBody : node.elseBody, new Context(context))
+      }
+
+      if(node.is('Throw')) {
+        const exception = yield* this.exec(node.exception)
+
+        yield node
+
+        throw new WollokException([...this.frameStack], exception)
+      }
+
+      if(node.is('Try')) {
+        yield node
+
+        let result: RuntimeObject | undefined
+        try {
+          result = yield* this.exec(node.body, new Context(context))
+        } catch(error) {
+          if(!(error instanceof WollokException)) throw error
+
+          const handler = node.catches.find(catcher => error.instance.module.inherits(catcher.parameterType.target()))
+
+          if(handler) result = yield* this.exec(handler.body, new Context(context, { [handler.parameter.name]: error.instance }))
+          else throw error
+        } finally {
+          yield* this.exec(node.always, new Context(context))
+        }
+
+        return result
+      }
+
+    } catch(error) {
+      if(error instanceof WollokException || error instanceof WollokReturn) throw error
+      else {
+        const module = this.environment.getNodeByFQN<Class>(error.message === 'Maximum call stack size exceeded'
+          ? 'wollok.lang.StackOverflowException'
+          : 'wollok.lang.EvaluationError'
+        )
+        throw new WollokException([...this.frameStack], new RuntimeObject(module, context, error))
+      }
+    } finally {
+      this.frameStack.pop()
+    }
+
+    throw new Error(`${node.kind} nodes can't be executed`)
+  }
+
+
+  *invoke(methodOrMessage: Method | Name | undefined, receiver: RuntimeObject, ...args: RuntimeObject[]): Generator<Node, RuntimeObject | undefined> {
+    const method = methodOrMessage instanceof Method ? methodOrMessage :
+      typeof methodOrMessage === 'string' ? receiver.module.lookupMethod(methodOrMessage, args.length) :
+      methodOrMessage
+
+    if (!method) return yield* this.invoke('messageNotUnderstood', receiver, yield* this.reify(methodOrMessage as string), yield* this.list(args))
+    if (method.isAbstract()) throw new Error(`Can't invoke abstract method ${method.parent().fullyQualifiedName()}.${method.name}/${method.parameters.length}`)
+    if (!method.matchesSignature(method.name, args.length)) throw new Error(`Wrong number of arguments (${args.length}) for method ${method.parent().fullyQualifiedName()}.${method.name}/${method.parameters.length}`)
+
+    if(method.body === 'native') {
+      const nativeFQN = `${method.parent().fullyQualifiedName()}.${method.name}`
+      const native = get<NativeFunction>(this.natives, nativeFQN)
+      if(!native) throw new Error(`Missing native ${nativeFQN}`)
+
+      this.frameStack.push(new Frame(method, new Context(receiver)))
+      try {
+        return yield* native.bind(this)(receiver, ...args)
+      } finally { this.frameStack.pop() }
+    } else {
+      let result: RuntimeObject | undefined
+      const locals: Record<Name, RuntimeObject> = {}
+      for(let index = 0; index < method.parameters.length; index++){
+        const { name, isVarArg } = method.parameters[index]
+        locals[name] = isVarArg ? yield* this.list(args.slice(index)) : args[index]
+      }
+
+      try {
+        result = yield* this.exec(method.body!, new Context(receiver, locals))
+      } catch(error) {
+        if(!(error instanceof WollokReturn)) throw error
+        else result = error.instance
+      }
+      return result
+    }
+  }
+
+  // ──────────────────────────────────────────────────────────────────────────────────────────────────────────────────
+  // INSTANCIATION
+  // ──────────────────────────────────────────────────────────────────────────────────────────────────────────────────
+
+  *reify(value: boolean | number | string | null): Generator<Node, RuntimeObject> {
+    if(typeof value === 'number') {
+      const stringValue = value.toFixed(DECIMAL_PRECISION)
+      const existing = this.rootContext.get(`N!${stringValue}`)
+      if(existing) return existing
+      const instance = new RuntimeObject(this.environment.getNodeByFQN('wollok.lang.Number'), this.rootContext, Number(stringValue))
+      this.rootContext.set(`N!${stringValue}`, instance)
+      return instance
+    }
+
+    if(typeof value === 'string'){
+      const existing = this.rootContext.get(`S!${value}`)
+      if(existing) return existing
+      const instance = new RuntimeObject(this.environment.getNodeByFQN('wollok.lang.String'), this.rootContext, value)
+      this.rootContext.set(`S!${value}`, instance)
+      return instance
+    }
+
+    if(typeof value === 'boolean'){
+      const existing = this.rootContext.get(`${value}`)
+      if(existing) return existing
+      const instance = new RuntimeObject(this.environment.getNodeByFQN('wollok.lang.Boolean'), this.rootContext, value)
+      this.rootContext.set(`${value}`, instance)
+      return instance
+    }
+
+    return this.rootContext.get(`${value}`)!
+  }
+
+  *list(value: RuntimeObject[]): Generator<Node, RuntimeObject> {
+    return new RuntimeObject(this.environment.getNodeByFQN('wollok.lang.List'), this.currentContext, value)
+  }
+
+  *set(value: RuntimeObject[]): Generator<Node, RuntimeObject> {
+    const result = new RuntimeObject(this.environment.getNodeByFQN('wollok.lang.Set'), this.currentContext, [])
+    for(const elem of value)
+      yield* this.invoke('add', result, elem)
+    return result
+  }
+
+  *instantiate(module: Module, locals: Record<Name, RuntimeObject> = {}): Generator<Node, RuntimeObject> {
+    const defaultFieldValues = module.defaultFieldValues()
+
+    const allFieldNames = [...defaultFieldValues.keys()].map(({ name }) => name)
+    for(const local of keys(locals))
+      if(!allFieldNames.includes(local))
+        throw new Error(`Can't instantiate ${module.fullyQualifiedName()} with value for unexistent field ${local}`)
+
+    const instance = new RuntimeObject(module, this.currentContext)
+
+    for(const [field, defaultValue] of defaultFieldValues) {
+      instance.set(field.name, field.name in locals ? locals[field.name] : defaultValue && this.exec(defaultValue, instance))
+    }
+
+    yield * this.invoke('initialize', instance)
+
+    if(!module.name || module.is('Describe'))
+      for (const [field] of defaultFieldValues)
+        instance.get(field.name)
+
+    return instance
   }
 }
