--- src/wre2/vm.ts
@@ -1,3 +1,0 @@
-// TODO: Implement
-
-export default {}
